review: общая функция обработки запроса #discussion_r1164691733
