+++ src/api/get-ingredients.js
@@ -1,0 +1,7 @@
+import request from "../utils/request";
+
+export default function getIngredients() {
+  return request("/ingredients")
+    .then((res) => res.data ?? Promise.reject())
+    .catch(() => Promise.reject("Ошибка при загрузке ингредиентов"));
+}
